+++ src/deps/ntars/base/utils.cpp
@@ -2,26 +2,26 @@
 
 namespace NTARS
 {
-    double meanSquaredError(const double y[], const double y_predicted[], uint32_t size)
+    float meanSquaredError(const float y[], const float y_predicted[], uint32_t size)
     {
-        double sum = std::inner_product(y, y + size, y_predicted, 0.0, std::plus<>(), [](double a, double b) {
+        float sum = std::inner_product(y, y + size, y_predicted, 0.0, std::plus<>(), [](float a, float b) {
             return std::pow(a - b, 2);
         });
         return sum / size;
     }
 
-    double meanAbsoluteError(const double y[], const double y_predicted[], uint32_t size)
+    float meanAbsoluteError(const float y[], const float y_predicted[], uint32_t size)
     {
-        double sum = std::inner_product(y, y + size, y_predicted, 0.0, std::plus<>(), [](double a, double b) {
+        float sum = std::inner_product(y, y + size, y_predicted, 0.0, std::plus<>(), [](float a, float b) {
             return std::abs(a - b);
         });
         return sum / size;
     }
 
-    double crossEntropyLoss(const double y[], double y_predicted[], uint32_t size)
+    float crossEntropyLoss(const float y[], float y_predicted[], uint32_t size)
     {
-        double sum = std::inner_product(y, y + size, y_predicted, 0.0, std::plus<>(), [](double a, double b) {
-            a += std::numeric_limits<double>().epsilon();
+        float sum = std::inner_product(y, y + size, y_predicted, 0.0, std::plus<>(), [](float a, float b) {
+            a += std::numeric_limits<float>().epsilon();
             return -(a * std::log(b) + (1 - a) * std::log(1 - b)); 
         });
         return sum / size;
