+++ src/deps/ntars/base/utils.hpp
@@ -8,9 +8,9 @@
 
 namespace NTARS
 {
-    double meanSquaredError(const double y[], const double y_predicted[], uint32_t size);
-    double meanAbsoluteError(const double y[], const double y_predicted[], uint32_t size);
-    double crossEntropyLoss(const double y[], double y_predicted[], uint32_t size);
+    float meanSquaredError(const float y[], const float y_predicted[], uint32_t size);
+    float meanAbsoluteError(const float y[], const float y_predicted[], uint32_t size);
+    float crossEntropyLoss(const float y[], float y_predicted[], uint32_t size);
 } // namespace NTARS
